Revert "wip link-menu"
