--- components/tree-view/link.js
@@ -1,3 +1,0 @@
-export default ({value}) => {
-  return <a target="_blank" href={value}>{value}</a>
-}
