+++ components/tree-view/menu.js
@@ -3,21 +3,47 @@ import { getState } from './state'
 import { hasChildren } from './analyze'
 import useClickOutside from './use-click-outside'
 
-export const MenuItem = ({ onClick, children }) => {
-  return <div><button onClick={onClick}>{children}</button></div>
-}
-
-export default ({ onClose, theme, children }) => {
+export default ({ onPickId, name, value, path, state, commandId, onSubmitForm, onClose, theme }) => {
   const ref = useRef(null)
   useClickOutside(ref, onClose)
+  const isArray = Array.isArray(value)
+  const viewType = getState(state)._viewType || 'tree'
+
+  const setViewType = viewType => {
+    onSubmitForm({
+      message: '_tree update',
+      commandId,
+      formData: {
+        path,
+        state: { ...state, _viewType: viewType, _expanded: true }
+      }
+    })
+    onClose()
+  }
+
+  const pickId = () => {
+    onPickId(name)
+    onClose()
+  }
+
+  const copyValue = () => {
+    window.Clipboard.writeText(JSON.stringify(value))
+  }
 
   return <div className="outer" ref={ref}>
     <div className="inner">
       {
-        children.map(item => {
-          return item 
+        ['tree', 'table', 'json'].map(key => {
+          if (key === viewType) {
+            return null
+          }
+          if (key === 'table' && !hasChildren(value)) {
+            return null
+          }
+          return <div key={key}><button onClick={() => setViewType(key)}>View as {key}</button></div>
         })
       }
+      <div><button onClick={pickId}>Paste into console</button></div>
     </div>
     <style jsx>{`
       div {
