+++ components/tree-view/table-view.js
@@ -1,10 +1,9 @@
 import { getCollectionPaths, displayPath, getAtPath, detectUrl } from './analyze'
-import Link from './link'
 
 const Summary = ({value}) => {
   if (typeof value === 'string') {
     if (detectUrl(value)) {
-      return <Link value={value} /> 
+      return <a target="_blank" href={value}>{value}</a>
     } else {
       return value
     }
